Move Rect to separate header
